Refactored and fixed so that the settings menu can be accessed from both the main menu and the ingame menu.

+++ librecraft/src/iGameState.h
@@ -6,7 +6,7 @@ public:
 
   virtual ~IGameState() {}
 
-  virtual void update(float timePassed) = 0;
+  virtual void update(double timePassed) = 0;
 
 };
